requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose
